pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "molbatch"
version = "0.2.0"
description = "Batch structure processing toolkit for PyMOL and ChimeraX"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "PerinMu" }
]
keywords = ["pymol", "chimerax", "structure", "batch", "visualization", "mmcif", "pdb"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
]
dependencies = [
  'tomli; python_version < "3.11"',
  'PyYAML>=6.0',
  'tomli-w>=1.0',
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[project.scripts]
molbatch = "molbatch.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
